+++ Core/Core.vcxproj.filters
@@ -523,6 +523,9 @@
     <ClCompile Include="HW\SimpleAudioDec.cpp">
       <Filter>HW</Filter>
     </ClCompile>
+    <ClCompile Include="MIPS\x86\JitSafeMem.cpp">
+      <Filter>MIPS\x86</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -965,6 +968,9 @@
     </ClInclude>
     <ClInclude Include="HW\SimpleAudioDec.h">
       <Filter>HW</Filter>
+    </ClInclude>
+    <ClInclude Include="MIPS\x86\JitSafeMem.h">
+      <Filter>MIPS\x86</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
